// src/types/generated/kusama/v9090.ts
export interface HrmpChannelId {
    sender: number
    receiver: number
}
// ...
export type ParaId = number
// ...
export const HrmpChannelId: sts.Type<HrmpChannelId> = sts.struct(() => {
    return  {
        sender: sts.number(),
        receiver: sts.number(),
    }
})

export const ParaId = sts.number()