// src/types/generated/kusama/v9100.ts
export type ParaId = number
// ...
export interface HrmpChannelId {
    sender: number
    receiver: number
}
// ...
export const HrmpChannelId: sts.Type<HrmpChannelId> = sts.struct(() => {
    return  {
        sender: sts.number(),
        receiver: sts.number(),
    }
})

export const Weight = sts.bigint()

export const OverweightIndex = sts.bigint()

export const ParaId = sts.number()