// src/types/generated/kusama/v9111.ts
export interface HrmpChannelId {
    sender: Id
    recipient: Id
}
// ...
export type Id = number
// ...
export const HrmpChannelId: sts.Type<HrmpChannelId> = sts.struct(() => {
    return  {
        sender: Id,
        recipient: Id,
    }
})
// ...
export const Id = sts.number()